+++ src/tool/cpp/code_writers.h
@@ -2,7 +2,7 @@
 
 namespace xlang
 {
-    void write_license(writer& w)
+    static void write_license(writer& w)
     {
         auto format = R"(// WARNING: Please don't edit this file. It was generated by C++/WinRT v%
 )";
@@ -10,7 +10,7 @@ namespace xlang
         w.write(format, XLANG_VERSION_STRING);
     }
 
-    void write_include_guard(writer& w)
+    static void write_include_guard(writer& w)
     {
         auto format = R"(#pragma once
 )";
@@ -18,7 +18,7 @@ namespace xlang
         w.write(format);
     }
 
-    void write_pch(writer& w)
+    static void write_pch(writer& w)
     {
         auto format = R"(#include "%"
 )";
@@ -29,7 +29,7 @@ namespace xlang
         }
     }
 
-    void write_impl_namespace(writer& w)
+    static void write_impl_namespace(writer& w)
     {
         auto format = R"(namespace winrt::impl
 {
@@ -38,21 +38,14 @@ namespace xlang
         w.write(format);
     }
 
-    void write_root_namespace(writer& w)
-    {
-        w.write(R"(namespace winrt
-{
-)");
-    }
-
-    void write_std_namespace(writer& w)
+    static void write_std_namespace(writer& w)
     {
         w.write(R"(namespace std
 {
 )");
     }
 
-    void write_type_namespace(writer& w, std::string_view const& ns)
+    static void write_type_namespace(writer& w, std::string_view const& ns)
     {
         auto format = R"(namespace winrt::@
 {
@@ -61,7 +54,7 @@ namespace xlang
         w.write(format, ns);
     }
 
-    void write_close_namespace(writer& w)
+    static void write_close_namespace(writer& w)
     {
         auto format = R"(}
 )";
@@ -69,7 +62,7 @@ namespace xlang
         w.write(format);
     }
 
-    void write_enum_field(writer& w, Field const& field)
+    static void write_enum_field(writer& w, Field const& field)
     {
         auto format = R"(        % = %,
 )";
@@ -80,7 +73,7 @@ namespace xlang
         }
     }
 
-    void write_enum(writer& w, TypeDef const& type)
+    static void write_enum(writer& w, TypeDef const& type)
     {
         auto format = R"(    enum class % : %
     {
@@ -91,7 +84,7 @@ namespace xlang
         w.write(format, type.TypeName(), fields.first.Signature().Type(), bind_each<write_enum_field>(fields));
     }
 
-    void write_forward(writer& w, TypeDef const& type)
+    static void write_forward(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
 
@@ -115,7 +108,7 @@ namespace xlang
         }
     }
 
-    void write_enum_flag(writer& w, TypeDef const& type)
+    static void write_enum_flag(writer& w, TypeDef const& type)
     {
         auto format = R"(    template<> struct is_enum_flag<@::%> : std::true_type
     {
@@ -128,7 +121,7 @@ namespace xlang
         }
     }
 
-    void write_category(writer& w, TypeDef const& type, std::string_view const& category)
+    static void write_category(writer& w, TypeDef const& type, std::string_view const& category)
     {
         auto format = R"(    template <> struct category<%>
     {
@@ -139,7 +132,7 @@ namespace xlang
         w.write(format, type, category);
     }
 
-    void write_name(writer& w, TypeDef const& type)
+    static void write_name(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <> struct name<@::%>
     {
@@ -153,7 +146,7 @@ namespace xlang
         w.write(format, type_namespace, type_name, type_namespace, type_name);
     }
 
-    void write_guid_value(writer& w, std::vector<FixedArgSig> const& args)
+    static void write_guid_value(writer& w, std::vector<FixedArgSig> const& args)
     {
         using std::get;
 
@@ -171,7 +164,7 @@ namespace xlang
             get<uint8_t>(get<ElemSig>(args[10].value).value));
     }
 
-    void write_guid(writer& w, TypeDef const& type)
+    static void write_guid(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <> struct guid_storage<%>
     {
@@ -189,7 +182,7 @@ namespace xlang
         w.write(format, type, bind<write_guid_value>(attribute.Value().FixedArgs()));
     }
 
-    void write_fast_version(writer& w, TypeDef const& type)
+    static void write_fast_version(writer& w, TypeDef const& type)
     {
         if (!is_fast_class(type))
         {
@@ -209,7 +202,7 @@ namespace xlang
             interfaces.back().name);
     }
 
-    void write_default_interface(writer& w, TypeDef const& type)
+    static void write_default_interface(writer& w, TypeDef const& type)
     {
         if (auto default_interface = get_default_interface(type))
         {
@@ -234,7 +227,7 @@ namespace xlang
         }
     }
 
-    void write_struct_category(writer& w, TypeDef const& type)
+    static void write_struct_category(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <> struct category<%>
     {
@@ -245,7 +238,7 @@ namespace xlang
         w.write(format, type, bind_list(", ", type.FieldList()));
     }
 
-    void write_array_size_name(writer& w, Param const& param)
+    static void write_array_size_name(writer& w, Param const& param)
     {
         if (w.param_names)
         {
@@ -253,7 +246,7 @@ namespace xlang
         }
     }
 
-    void write_abi_params(writer& w, method_signature const& method_signature)
+    static void write_abi_params(writer& w, method_signature const& method_signature)
     {
         w.abi_types = true;
         separator s{ w };
@@ -371,7 +364,7 @@ namespace xlang
         return false;
     }
 
-    void write_abi_args(writer& w, method_signature const& method_signature)
+    static void write_abi_args(writer& w, method_signature const& method_signature)
     {
         separator s{ w };
 
@@ -455,17 +448,7 @@ namespace xlang
         }
     }
 
-    void write_abi_declaration(writer& w, MethodDef const& method)
-    {
-        auto format = R"(        virtual int32_t WINRT_CALL %(%) noexcept = 0;
-)";
-
-        w.param_names = false;
-        method_signature signature{ method };
-        w.write(format, get_abi_name(method), bind<write_abi_params>(signature));
-    }
-
-    void write_fast_abi(writer& w, TypeDef const& type)
+    static void write_fast_abi(writer& w, TypeDef const& type)
     {
         if (!is_fast_class(type))
         {
@@ -501,7 +484,7 @@ namespace xlang
 )");
     }
 
-    void write_interface_abi(writer& w, TypeDef const& type)
+    static void write_interface_abi(writer& w, TypeDef const& type)
     {
         auto guard{ w.push_generic_params(type.GenericParam()) };
 
@@ -527,7 +510,7 @@ namespace xlang
 )");
     }
 
-    void write_delegate_abi(writer& w, TypeDef const& type)
+    static void write_delegate_abi(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <> struct abi<@::%>
     {
@@ -548,12 +531,12 @@ namespace xlang
             bind<write_abi_params>(signature));
     }
 
-    void write_field_abi(writer& w, Field const& field)
+    static void write_field_abi(writer& w, Field const& field)
     {
         w.write("        % %;\n", get_field_abi(w, field), field.Name());
     }
 
-    void write_struct_abi(writer& w, TypeDef const& type)
+    static void write_struct_abi(writer& w, TypeDef const& type)
     {
         w.abi_types = true;
 
@@ -579,7 +562,7 @@ namespace xlang
 
     }
 
-    void write_consume_params(writer& w, method_signature const& signature)
+    static void write_consume_params(writer& w, method_signature const& signature)
     {
         separator s{ w };
 
@@ -639,7 +622,7 @@ namespace xlang
         }
     }
 
-    void write_implementation_params(writer& w, method_signature const& method_signature)
+    static void write_implementation_params(writer& w, method_signature const& method_signature)
     {
         separator s{ w };
 
@@ -699,7 +682,7 @@ namespace xlang
         }
     }
 
-    void write_consume_declaration(writer& w, MethodDef const& method)
+    static void write_consume_declaration(writer& w, MethodDef const& method)
     {
         method_signature signature{ method };
         w.async_types = is_async(method, signature);
@@ -729,7 +712,7 @@ namespace xlang
         }
     }
 
-    void write_consume_return_type(writer& w, method_signature const& signature)
+    static void write_consume_return_type(writer& w, method_signature const& signature)
     {
         if (!signature.return_signature())
         {
@@ -748,7 +731,7 @@ namespace xlang
         }
     }
 
-    void write_consume_return_statement(writer& w, method_signature const& signature)
+    static void write_consume_return_statement(writer& w, method_signature const& signature)
     {
         if (!signature.return_signature())
         {
@@ -765,7 +748,7 @@ namespace xlang
         }
     }
 
-    void write_consume_args(writer& w, method_signature const& signature)
+    static void write_consume_args(writer& w, method_signature const& signature)
     {
         separator s{ w };
 
@@ -776,7 +759,7 @@ namespace xlang
         }
     }
 
-    void write_consume_definitions(writer& w, TypeDef const& type)
+    static void write_consume_definitions(writer& w, TypeDef const& type)
     {
         auto guard{ w.push_generic_params(type.GenericParam()) };
         auto type_name = type.TypeName();
@@ -841,7 +824,7 @@ namespace xlang
         }
     }
 
-    void write_consume_extensions(writer& w, TypeDef const& type)
+    static void write_consume_extensions(writer& w, TypeDef const& type)
     {
         if (type.TypeName() == "IBindableIterator" && type.TypeNamespace() == "Windows.UI.Xaml.Interop")
         {
@@ -875,7 +858,7 @@ namespace xlang
         }
     }
 
-    void write_consume(writer& w, TypeDef const& type)
+    static void write_consume(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <typename D>
     struct consume_%
@@ -902,13 +885,13 @@ namespace xlang
             impl_name);
     }
 
-    void write_produce_params(writer& w, method_signature const& signature)
+    static void write_produce_params(writer& w, method_signature const& signature)
     {
         w.param_names = true;
         write_abi_params(w, signature);
     }
 
-    void write_produce_cleanup_param(writer& w, TypeSig const& signature, std::string_view const& param_name, bool out)
+    static void write_produce_cleanup_param(writer& w, TypeSig const& signature, std::string_view const& param_name, bool out)
     {
         if (signature.is_szarray())
         {
@@ -983,7 +966,7 @@ namespace xlang
         }
     }
 
-    void write_produce_cleanup(writer& w, method_signature const& method_signature)
+    static void write_produce_cleanup(writer& w, method_signature const& method_signature)
     {
         for (auto&&[param, param_signature] : method_signature.params())
         {
@@ -1001,7 +984,7 @@ namespace xlang
         }
     }
 
-    void write_produce_args(writer& w, method_signature const& method_signature)
+    static void write_produce_args(writer& w, method_signature const& method_signature)
     {
         w.abi_types = false;
         separator s{ w };
@@ -1078,7 +1061,7 @@ namespace xlang
         }
     }
 
-    void write_produce_upcall(writer& w, MethodDef const& method, method_signature const& method_signature)
+    static void write_produce_upcall(writer& w, MethodDef const& method, method_signature const& method_signature)
     {
         w.abi_types = false;
 
@@ -1121,7 +1104,7 @@ namespace xlang
         }
     }
 
-    void write_delegate_upcall(writer& w, method_signature const& method_signature)
+    static void write_delegate_upcall(writer& w, method_signature const& method_signature)
     {
         w.abi_types = false;
 
@@ -1151,7 +1134,7 @@ namespace xlang
         }
     }
 
-    void write_produce_method(writer& w, MethodDef const& method)
+    static void write_produce_method(writer& w, MethodDef const& method)
     {
         std::string_view format;
 
@@ -1190,7 +1173,7 @@ namespace xlang
             bind<write_produce_upcall>(method, signature));
     }
 
-    void write_fast_produce_methods(writer& w, TypeDef const& type)
+    static void write_fast_produce_methods(writer& w, TypeDef const& type)
     {
         for (auto&& info : get_fast_interfaces(w, type))
         {
@@ -1201,7 +1184,7 @@ namespace xlang
         }
     }
 
-    void write_produce(writer& w, TypeDef const& type)
+    static void write_produce(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <typename D>
     struct produce<D, %> : produce_base<D, %>
@@ -1217,7 +1200,7 @@ namespace xlang
             bind_each<write_produce_method>(type.MethodList()));
     }
 
-    void write_fast_produce(writer& w, TypeDef const& type)
+    static void write_fast_produce(writer& w, TypeDef const& type)
     {
         if (!is_fast_class(type))
         {
@@ -1238,7 +1221,7 @@ namespace xlang
             bind<write_fast_produce_methods>(type));
     }
 
-    void write_dispatch_overridable_method(writer& w, MethodDef const& method)
+    static void write_dispatch_overridable_method(writer& w, MethodDef const& method)
     {
         auto format = R"(    % %(%)
     {
@@ -1263,7 +1246,7 @@ namespace xlang
             bind<write_consume_args>(signature));
     }
 
-    void write_dispatch_overridable(writer& w, TypeDef const& class_type)
+    static void write_dispatch_overridable(writer& w, TypeDef const& class_type)
     {
         auto format = R"(template <typename T, typename D>
 struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
@@ -1283,7 +1266,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_interface_override_method(writer& w, MethodDef const& method, std::string_view const& interface_name)
+    static void write_interface_override_method(writer& w, MethodDef const& method, std::string_view const& interface_name)
     {
         auto format = R"(template <typename D> % %T<D>::%(%) const
 {
@@ -1304,7 +1287,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
             bind<write_consume_args>(signature));
     }
 
-    void write_interface_override_methods(writer& w, TypeDef const& class_type)
+    static void write_interface_override_methods(writer& w, TypeDef const& class_type)
     {
         for (auto&&[interface_name, info] : get_interfaces(w, class_type))
         {
@@ -1317,7 +1300,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         };
     }
 
-    void write_class_override_implements(writer& w, std::map<std::string, interface_info> const& interfaces)
+    static void write_class_override_implements(writer& w, std::map<std::string, interface_info> const& interfaces)
     {
         bool found{};
 
@@ -1336,7 +1319,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_class_override_requires(writer& w, std::map<std::string, interface_info> const& interfaces)
+    static void write_class_override_requires(writer& w, std::map<std::string, interface_info> const& interfaces)
     {
         bool found{};
 
@@ -1350,7 +1333,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_class_override_defaults(writer& w, std::map<std::string, interface_info> const& interfaces)
+    static void write_class_override_defaults(writer& w, std::map<std::string, interface_info> const& interfaces)
     {
         bool first{ true };
 
@@ -1373,7 +1356,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_component_override_defaults(writer& w, TypeDef const& type)
+    static void write_component_override_defaults(writer& w, TypeDef const& type)
     {
         std::vector<std::string> interfaces;
 
@@ -1414,7 +1397,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_class_override_bases(writer& w, TypeDef const& type)
+    static void write_class_override_bases(writer& w, TypeDef const& type)
     {
         for (auto&& base : get_bases(type))
         {
@@ -1422,7 +1405,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_class_override_constructors(writer& w, std::string_view const& type_name, std::vector<factory_type> const& factories)
+    static void write_class_override_constructors(writer& w, std::string_view const& type_name, std::vector<factory_type> const& factories)
     {
         auto format = R"(    %T(%)
     {
@@ -1455,7 +1438,7 @@ struct WINRT_EBO produce_dispatch_to_overridable<T, D, %>
         }
     }
 
-    void write_interface_override(writer& w, TypeDef const& type)
+    static void write_interface_override(writer& w, TypeDef const& type)
     {
         auto format = R"(template <typename D>
 class %T
@@ -1486,7 +1469,7 @@ public:
         }
     }
 
-    void write_class_override(writer& w, TypeDef const& type)
+    static void write_class_override(writer& w, TypeDef const& type)
     {
         auto factories = get_factories(type);
         bool has_composable_factories{};
@@ -1531,7 +1514,7 @@ public:
             bind<write_class_override_constructors>(type_name, factories));
     }
 
-    void write_interface_requires(writer& w, TypeDef const& type)
+    static void write_interface_requires(writer& w, TypeDef const& type)
     {
         auto interfaces = get_interfaces(w, type);
 
@@ -1550,7 +1533,7 @@ public:
         w.write('>');
     }
 
-    void write_interface_usings(writer& w, TypeDef const& type)
+    static void write_interface_usings(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
         auto interfaces_plus_self = get_interfaces(w, type);
@@ -1582,7 +1565,7 @@ public:
         }
     }
 
-    void write_class_usings(writer& w, TypeDef const& type)
+    static void write_class_usings(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
         auto default_interface = get_default_interface(type);
@@ -1634,7 +1617,7 @@ public:
 
     }
 
-    void write_interface(writer& w, TypeDef const& type)
+    static void write_interface(writer& w, TypeDef const& type)
     {
         auto format = R"(    struct WINRT_EBO % :
         Windows::Foundation::IInspectable,
@@ -1658,7 +1641,7 @@ public:
             bind<write_interface_usings>(type));
     }
 
-    void write_delegate(writer& w, TypeDef const& type)
+    static void write_delegate(writer& w, TypeDef const& type)
     {
         auto format = R"(    struct % : Windows::Foundation::IUnknown
     {
@@ -1688,7 +1671,7 @@ public:
             bind<write_consume_params>(signature));
     }
 
-    void write_delegate_implementation(writer& w, TypeDef const& type)
+    static void write_delegate_implementation(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <> struct delegate<@::%>
     {
@@ -1728,7 +1711,7 @@ public:
             "");
     }
 
-    void write_delegate_definition(writer& w, TypeDef const& type)
+    static void write_delegate_definition(writer& w, TypeDef const& type)
     {
         auto format = R"(    template <typename L> %::%(L handler) :
         %(impl::make_delegate<%>(std::forward<L>(handler)))
@@ -1786,14 +1769,14 @@ public:
             bind<write_consume_return_statement>(signature));
     }
 
-    void write_struct_field(writer& w, std::pair<std::string_view, std::string> const& field)
+    static void write_struct_field(writer& w, std::pair<std::string_view, std::string> const& field)
     {
         w.write("        @ %;\n",
             field.second,
             field.first);
     }
 
-    void write_struct_equality(writer& w, std::vector<std::pair<std::string_view, std::string>> const& fields)
+    static void write_struct_equality(writer& w, std::vector<std::pair<std::string_view, std::string>> const& fields)
     {
         for (size_t i = 0; i != fields.size(); ++i)
         {
@@ -1806,7 +1789,7 @@ public:
         }
     }
 
-    void write_structs(writer& w, std::vector<TypeDef> const& types)
+    static void write_structs(writer& w, std::vector<TypeDef> const& types)
     {
         auto format = R"(    struct %
     {
@@ -1904,7 +1887,7 @@ public:
         }
     }
 
-    void write_slow_class_requires(writer& w, TypeDef const& type)
+    static void write_slow_class_requires(writer& w, TypeDef const& type)
     {
         bool first{ true };
 
@@ -1928,7 +1911,7 @@ public:
         }
     }
 
-    void write_fast_class_requires(writer& w, TypeDef const& type)
+    static void write_fast_class_requires(writer& w, TypeDef const& type)
     {
         bool first{ true };
 
@@ -1952,7 +1935,7 @@ public:
         }
     }
 
-    void write_component_class_base(writer& w, TypeDef const& type)
+    static void write_component_class_base(writer& w, TypeDef const& type)
     {
         bool first{ true };
 
@@ -1978,7 +1961,7 @@ public:
         }
     }
 
-    void write_class_base(writer& w, TypeDef const& type)
+    static void write_class_base(writer& w, TypeDef const& type)
     {
         bool first{ true };
 
@@ -1999,7 +1982,7 @@ public:
         }
     }
 
-    void write_constructor_declarations(writer& w, TypeDef const& type, std::vector<factory_type> const& factories)
+    static void write_constructor_declarations(writer& w, TypeDef const& type, std::vector<factory_type> const& factories)
     {
         auto type_name = type.TypeName();
 
@@ -2039,7 +2022,7 @@ public:
         }
     }
 
-    void write_constructor_definition(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
+    static void write_constructor_definition(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
     {
         method_signature signature{ method };
 
@@ -2060,7 +2043,7 @@ public:
             bind<write_consume_args>(signature));
     }
 
-    void write_composable_constructor_definition(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
+    static void write_composable_constructor_definition(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
     {
         method_signature signature{ method };
         auto& params = signature.params();
@@ -2092,7 +2075,7 @@ public:
     }
 
 
-    void write_static_declaration(writer& w, factory_type const& factory)
+    static void write_static_declaration(writer& w, factory_type const& factory)
     {
         if (!factory.statics)
         {
@@ -2128,7 +2111,7 @@ public:
         }
     }
 
-    void write_static_definitions(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
+    static void write_static_definitions(writer& w, MethodDef const& method, std::string_view const& type_name, TypeDef const& factory)
     {
         method_signature signature{ method };
         auto method_name = get_name(method);
@@ -2175,7 +2158,7 @@ public:
         }
     }
 
-    void write_class_definitions(writer& w, TypeDef const& type)
+    static void write_class_definitions(writer& w, TypeDef const& type)
     {
         if (settings.component_opt && settings.filter.includes(type))
         {
@@ -2283,7 +2266,7 @@ public:
         }
     }
 
-    void write_fast_declarations(writer& w, TypeDef const& type)
+    static void write_fast_declarations(writer& w, TypeDef const& type)
     {
         for (auto&& info : get_fast_interfaces(w, type))
         {
@@ -2291,7 +2274,7 @@ public:
         }
     }
 
-    void write_slow_class(writer& w, TypeDef const& type, coded_index<TypeDefOrRef> const& base_type)
+    static void write_slow_class(writer& w, TypeDef const& type, coded_index<TypeDefOrRef> const& base_type)
     {
         auto type_name = type.TypeName();
         auto factories = get_factories(type);
@@ -2316,7 +2299,7 @@ public:
             bind_each<write_static_declaration>(factories));
     }
 
-    void write_fast_class(writer& w, TypeDef const& type)
+    static void write_fast_class(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
         auto factories = get_factories(type);
@@ -2352,7 +2335,7 @@ public:
             bind<write_fast_declarations>(type));
     }
 
-    void write_static_class(writer& w, TypeDef const& type)
+    static void write_static_class(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
         auto factories = get_factories(type);
@@ -2369,7 +2352,7 @@ public:
             bind_each<write_static_declaration>(factories));
     }
 
-    void write_class(writer& w, TypeDef const& type)
+    static void write_class(writer& w, TypeDef const& type)
     {
         if (auto default_interface = get_default_interface(type))
         {
@@ -2388,7 +2371,7 @@ public:
         }
     }
 
-    void write_std_hash(writer& w, TypeDef const& type)
+    static void write_std_hash(writer& w, TypeDef const& type)
     {
         auto type_name = type.TypeName();
         auto type_namespace = type.TypeNamespace();
@@ -2400,7 +2383,7 @@ public:
             type_name);
     }
 
-    void write_namespace_special(writer& w, std::string_view const& namespace_name, cache const& c)
+    static void write_namespace_special(writer& w, std::string_view const& namespace_name, cache const& c)
     {
         if (namespace_name == "Windows.Foundation")
         {
@@ -2481,1025 +2464,5 @@ public:
         {
             w.write(strings::base_resume_foreground);
         }
-    }
-
-    void write_base_fragment(writer& w, std::string_view const& value)
-    {
-        for (auto&& c : value)
-        {
-            if (c == '`')
-            {
-                w.write(settings.root);
-            }
-            else
-            {
-                w.write(c);
-            }
-        }
-    }
-
-    void write_component_include(writer& w, TypeDef const& type)
-    {
-        if (!has_factory_members(type))
-        {
-            return;
-        }
-
-        if (settings.component_opt)
-        {
-            auto format = R"(void* winrt_make_%();
-)";
-
-            w.write(format, get_impl_name(type.TypeNamespace(), type.TypeName()));
-        }
-        else
-        {
-            auto format = R"(#include "%.h"
-)";
-
-            w.write(format, get_component_filename(type));
-        }
-    }
-
-    void write_component_activation(writer& w, TypeDef const& type)
-    {
-        if (!has_factory_members(type))
-        {
-            return;
-        }
-
-        auto type_name = type.TypeName();
-        auto type_namespace = type.TypeNamespace();
-        auto impl_name = get_impl_name(type_namespace, type_name);
-
-        if (settings.component_opt)
-        {
-            auto format = R"(
-    if (requal(name, L"%.%"))
-    {
-        return winrt_make_%();
-    }
-)";
-
-            w.write(format,
-                type_namespace,
-                type_name,
-                impl_name);
-        }
-        else
-        {
-            auto format = R"(
-    if (requal(name, L"%.%"))
-    {
-        return winrt::detach_abi(winrt::make<winrt::@::factory_implementation::%>());
-    }
-)";
-
-            w.write(format,
-                type_namespace,
-                type_name,
-                type_namespace,
-                type_name);
-        }
-    }
-
-    void write_module_g_cpp(writer& w, std::vector<TypeDef> const& classes)
-    {
-        auto format = R"(#include "winrt/base.h"
-%
-bool WINRT_CALL %_can_unload_now() noexcept
-{
-    if (winrt::get_module_lock())
-    {
-        return false;
-    }
-
-    winrt::clear_factory_cache();
-    return true;
-}
-
-void* WINRT_CALL %_get_activation_factory(std::wstring_view const& name)
-{
-    auto requal = [](std::wstring_view const& left, std::wstring_view const& right) noexcept
-    {
-        return std::equal(left.rbegin(), left.rend(), right.rbegin(), right.rend());
-    };
-%
-    return nullptr;
-}
-
-int32_t WINRT_CALL WINRT_CanUnloadNow() noexcept
-{
-#ifdef _WRL_MODULE_H_
-    if (!::Microsoft::WRL::Module<::Microsoft::WRL::InProc>::GetModule().Terminate())
-    {
-        return 1;
-    }
-#endif
-
-    return %_can_unload_now() ? 0 : 1;
-}
-
-int32_t WINRT_CALL WINRT_GetActivationFactory(void* classId, void** factory) noexcept
-{
-    try
-    {
-        uint32_t length{};
-        wchar_t const* const buffer = WINRT_WindowsGetStringRawBuffer(classId, &length);
-        std::wstring_view const name{ buffer, length };
-
-        *factory = %_get_activation_factory(name);
-
-        if (*factory)
-        {
-            return 0;
-        }
-
-#ifdef _WRL_MODULE_H_
-        return ::Microsoft::WRL::Module<::Microsoft::WRL::InProc>::GetModule().GetActivationFactory(static_cast<HSTRING>(classId), reinterpret_cast<::IActivationFactory**>(factory));
-#else
-        return winrt::hresult_class_not_available(name).to_abi();
-#endif
-    }
-    catch (...) { return winrt::to_hresult(); }
-}
-)";
-
-        w.write(format,
-            bind_each<write_component_include>(classes),
-            settings.component_lib,
-            settings.component_lib,
-            bind_each<write_component_activation>(classes),
-            settings.component_lib,
-            settings.component_lib);
-    }
-
-    void write_component_interfaces(writer& w, TypeDef const& type)
-    {
-        if (is_fast_class(type))
-        {
-            w.write(", fast_interface<@::%>", type.TypeNamespace(), type.TypeName());
-
-            for (auto&&[interface_name, info] : get_interfaces(w, type))
-            {
-                if (!is_exclusive(info.type) && !info.base)
-                {
-                    w.write(", @", interface_name);
-                }
-            }
-        }
-        else
-        {
-            for (auto&&[interface_name, info] : get_interfaces(w, type))
-            {
-                if (!info.base)
-                {
-                    w.write(", @", interface_name);
-                }
-            }
-        }
-
-        if (has_composable_constructors(type))
-        {
-            w.write(", composable");
-        }
-
-        auto base_type = get_base_class(type);
-
-        if (!base_type)
-        {
-            return;
-        }
-
-        if (settings.filter.includes(base_type))
-        {
-            return;
-        }
-
-        w.write(", composing");
-
-        for (auto&&[interface_name, info] : get_interfaces(w, base_type))
-        {
-            if (info.overridable)
-            {
-                w.write(", @", interface_name);
-            }
-        }
-    }
-
-    void write_component_composable_forwarder(writer& w, MethodDef const& method)
-    {
-        auto format = R"(        % %(%)
-        {
-            return impl::composable_factory<T>::template CreateInstance<%>(%);
-        }
-)";
-
-        method_signature signature{ method };
-        method_signature reordered_method = signature;
-        auto&& params = reordered_method.params();
-        std::rotate(params.begin(), params.end() - 2, params.end());
-        w.param_names = true;
-
-        w.write(format,
-            signature.return_signature(),
-            get_name(method),
-            bind<write_implementation_params>(signature),
-            signature.return_signature(),
-            bind<write_consume_args>(reordered_method));
-    }
-
-    void write_component_constructor_forwarder(writer& w, MethodDef const& method)
-    {
-        auto format = R"(        % %(%)
-        {
-            return make<T>(%);
-        }
-)";
-
-        method_signature signature{ method };
-        w.param_names = true;
-
-        w.write(format,
-            signature.return_signature(),
-            get_name(method),
-            bind<write_implementation_params>(signature),
-            bind<write_consume_args>(signature));
-    }
-
-        void write_component_static_forwarder(writer& w, MethodDef const& method)
-    {
-        auto format = R"(        % %(%)
-        {
-            return T::%(%);
-        }
-)";
-
-        method_signature signature{ method };
-        w.param_names = true;
-
-        w.write(format,
-            signature.return_signature(),
-            get_name(method),
-            bind<write_implementation_params>(signature),
-            get_name(method),
-            bind<write_consume_args>(signature));
-    }
-
-    void write_component_forwarders(writer& w, std::vector<factory_type> const& factories)
-    {
-        bool default_constructor{};
-
-        for (auto&& factory : factories)
-        {
-            if (factory.activatable)
-            {
-                if (!factory.type)
-                {
-                    default_constructor = true;
-
-                    w.write(R"(        Windows::Foundation::IInspectable ActivateInstance() const
-        {
-            return make<T>();
-        }
-)");
-                }
-                else
-                {
-                    w.write_each<write_component_constructor_forwarder>(factory.type.MethodList());
-                }
-            }
-            else if (factory.statics)
-            {
-                w.write_each<write_component_static_forwarder>(factory.type.MethodList());
-            }
-            else if (factory.composable)
-            {
-                w.write_each<write_component_composable_forwarder>(factory.type.MethodList());
-            }
-        }
-
-        if (!default_constructor)
-        {
-            w.write(R"(        [[noreturn]] Windows::Foundation::IInspectable ActivateInstance() const
-        {
-            throw hresult_not_implemented();
-        }
-)");
-        }
-    }
-
-    void write_component_factory_interfaces(writer& w, std::vector<factory_type> const& factories)
-    {
-        for (auto&& factory : factories)
-        {
-            if (!factory.type)
-            {
-                continue;
-            }
-            
-            w.write(", %", factory.type);
-        }
-    }
-
-    void write_component_g_cpp(writer& w, TypeDef const& type)
-    {
-        auto type_name = type.TypeName();
-        auto type_namespace = type.TypeNamespace();
-        auto impl_name = get_impl_name(type_namespace, type_name);
-
-        if (has_factory_members(type))
-        {
-            auto format = R"(
-void* winrt_make_%()
-{
-    return winrt::detach_abi(winrt::make<winrt::@::factory_implementation::%>());
-}
-)";
-
-            w.write(format,
-                impl_name,
-                type_namespace,
-                type_name);
-        }
-
-        if (!settings.component_opt)
-        {
-            return;
-        }
-
-        write_type_namespace(w, type_namespace);
-
-        for (auto&& factory : get_factories(type))
-        {
-            if (factory.activatable)
-            {
-                if (!factory.type)
-                {
-                    auto format = R"(    %::%() :
-        %(make<@::implementation::%>())
-    {
-    }
-)";
-
-                    w.write(format,
-                        type_name,
-                        type_name,
-                        type_name,
-                        type_namespace,
-                        type_name);
-                }
-                else
-                {
-                    for (auto&& method : factory.type.MethodList())
-                    {
-                        method_signature signature{ method };
-
-                        auto format = R"(    %::%(%) :
-        %(make<@::implementation::%>(%))
-    {
-    }
-)";
-
-                        w.write(format,
-                            type_name,
-                            type_name,
-                            bind<write_consume_params>(signature),
-                            type_name,
-                            type_namespace,
-                            type_name,
-                            bind<write_consume_args>(signature));
-                    }
-                }
-            }
-            else if (factory.composable && factory.visible)
-            {
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-                    auto& params = signature.params();
-                    params.pop_back();
-                    params.pop_back();
-
-                    auto format = R"(    %::%(%) :
-        %(make<@::implementation::%>(%))
-    {
-    }
-)";
-
-                    w.write(format,
-                        type_name,
-                        type_name,
-                        bind<write_consume_params>(signature),
-                        type_name,
-                        type_namespace,
-                        type_name,
-                        bind<write_consume_args>(signature));
-                }
-            }
-            else if (factory.statics)
-            {
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-                    auto method_name = get_name(method);
-                    w.async_types = is_async(method, signature);
-
-                    if (is_add_overload(method) || is_remove_overload(method))
-                    {
-                        auto format = R"(    % %::%(%)
-    {
-        auto f = make<winrt::@::factory_implementation::%>().as<%>();
-        return f.%(%);
-    }
-)";
-
-
-                        w.write(format,
-                            signature.return_signature(),
-                            type_name,
-                            method_name,
-                            bind<write_consume_params>(signature),
-                            type_namespace,
-                            type_name,
-                            factory.type,
-                            method_name,
-                            bind<write_consume_args>(signature));
-                    }
-                    else
-                    {
-                        auto format = R"(    % %::%(%)
-    {
-        return @::implementation::%::%(%);
-    }
-)";
-
-
-                        w.write(format,
-                            signature.return_signature(),
-                            type_name,
-                            method_name,
-                            bind<write_consume_params>(signature),
-                            type_namespace,
-                            type_name,
-                            method_name,
-                            bind<write_consume_args>(signature));
-                    }
-
-                    if (is_add_overload(method))
-                    {
-                        auto format = R"(    %::%_revoker %::%(auto_revoke_t, %)
-    {
-        auto f = make<winrt::@::factory_implementation::%>().as<%>();
-        return { f, f.%(%) };
-    }
-)";
-
-                        w.write(format,
-                            type_name,
-                            method_name,
-                            type_name,
-                            method_name,
-                            bind<write_consume_params>(signature),
-                            type_namespace,
-                            type_name,
-                            factory.type,
-                            method_name,
-                            bind<write_consume_args>(signature));
-                    }
-                }
-            }
-        }
-
-        if (!is_fast_class(type))
-        {
-            write_close_namespace(w);
-            return;
-        }
-
-        for (auto&& info : get_fast_interfaces(w, type))
-        {
-            for (auto&& method : info.methods)
-            {
-                auto method_name = get_name(method);
-                method_signature signature{ method };
-                w.async_types = is_async(method, signature);
-
-                std::string_view format;
-
-                if (is_noexcept(method))
-                {
-                    format = R"(    % %::%(%) const noexcept
-    {
-        return get_self<@::implementation::%>(*this)->%(%);
-    }
-)";
-                }
-                else
-                {
-                    format = R"(    % %::%(%) const
-    {
-        return get_self<@::implementation::%>(*this)->%(%);
-    }
-)";
-                }
-
-                w.write(format,
-                    signature.return_signature(),
-                    type_name,
-                    method_name,
-                    bind<write_consume_params>(signature),
-                    type_namespace,
-                    type_name,
-                    method_name,
-                    bind<write_consume_args>(signature));
-
-                if (is_add_overload(method))
-                {
-                    format = R"(    %::%_revoker %::%(auto_revoke_t, %) const
-    {
-        return impl::make_event_revoker<%, %_revoker>(this, %(%));
-    }
-)";
-
-                    w.write(format,
-                        type_name,
-                        method_name,
-                        type_name,
-                        method_name,
-                        bind<write_consume_params>(signature),
-                        method_name,
-                        method_name,
-                        bind<write_consume_args>(signature));
-                }
-            }
-        }
-
-        write_close_namespace(w);
-    }
-
-    void write_component_override_dispatch_base(writer& w, TypeDef const& type)
-    {
-        if (!is_composable(type))
-        {
-            return;
-        }
-
-        std::string interfaces;
-
-        for (auto&& [name, info] : get_interfaces(w, type))
-        {
-            if (!info.overridable)
-            {
-                continue;
-            }
-
-            interfaces += ", ";
-            interfaces += name;
-        }
-
-        if (interfaces.empty())
-        {
-            return;
-        }
-
-        auto format = R"(
-    protected:
-        using dispatch = impl::dispatch_to_overridable<D@>;
-        auto overridable() noexcept { return dispatch::overridable(static_cast<D&>(*this)); }
-        )";
-
-        w.write(format, interfaces);
-    }
-
-    void write_component_class_override_constructors(writer& w, TypeDef const& type)
-    {
-        auto base_type = get_base_class(type);
-
-        if (!base_type)
-        {
-            return;
-        }
-
-        if (settings.filter.includes(base_type))
-        {
-            return;
-        }
-
-        auto type_name = type.TypeName();
-
-        for (auto&& factory : get_factories(base_type))
-        {
-            if (!factory.composable)
-            {
-                continue;
-            }
-
-            for (auto&& method : factory.type.MethodList())
-            {
-                method_signature signature{ method };
-                auto& params = signature.params();
-                params.resize(params.size() - 2);
-
-                auto format = R"(
-    %_base(%)
-    {
-        impl::call_factory<%, %>([&](auto&& f) { f.%(%%*this, this->m_inner); });
-    }
-                )";
-
-                w.write(format,
-                    type_name,
-                    bind<write_consume_params>(signature),
-                    base_type,
-                    factory.type,
-                    get_name(method),
-                    bind<write_consume_args>(signature),
-                    params.empty() ? "" : ", ");
-            }
-        }
-    }
-
-    void write_component_g_h(writer& w, TypeDef const& type)
-    {
-        auto type_name = type.TypeName();
-        auto type_namespace = type.TypeNamespace();
-        auto interfaces = get_interfaces(w, type);
-        auto factories = get_factories(type);
-        bool const non_static = !empty(type.InterfaceImpl());
-
-        if (non_static)
-        {
-            auto format = R"(namespace winrt::@::implementation
-{
-    template <typename D%, typename... I>
-    struct WINRT_EBO %_base : implements<D%%, %I...>%%%
-    {
-        using base_type = %_base;
-        using class_type = @::%;
-        using implements_type = typename %_base::implements_type;
-        using implements_type::implements_type;
-        %
-        operator impl::producer_ref<class_type> const() const noexcept
-        {
-            return { to_abi<default_interface<class_type>>(this) };
-        }
-
-        hstring GetRuntimeClassName() const
-        {
-            return L"%.%";
-        }
-    %%};
-}
-)";
-
-            auto base_type = get_base_class(type);
-            std::string composable_base_name;
-            std::string base_type_parameter;
-            std::string base_type_argument;
-            std::string no_module_lock;
-            bool external_base_type{};
-
-            if (base_type)
-            {
-                external_base_type = !settings.filter.includes(base_type);
-
-                if (external_base_type)
-                {
-                    composable_base_name = w.write_temp("using composable_base = %;", base_type);
-                }
-                else
-                {
-                    base_type_parameter = ", typename B";
-                    base_type_argument = ", B";
-                    no_module_lock = "no_module_lock, ";
-                }
-            }
-
-            w.write(format,
-                type_namespace,
-                base_type_parameter,
-                type_name,
-                bind<write_component_interfaces>(type),
-                base_type_argument,
-                no_module_lock,
-                "",
-                bind<write_component_class_base>(type),
-                bind<write_component_override_defaults>(type),
-                type_name,
-                type_namespace,
-                type_name,
-                type_name,
-                composable_base_name,
-                type_namespace,
-                type_name,
-                bind<write_component_class_override_constructors>(type),
-                bind<write_component_override_dispatch_base>(type));
-        }
-
-        if (has_factory_members(type))
-        {
-            auto format = R"(namespace winrt::@::factory_implementation
-{
-    template <typename D, typename T, typename... I>
-    struct WINRT_EBO %T : implements<D, Windows::Foundation::IActivationFactory%, I...>
-    {
-        using instance_type = @::%;
-
-        hstring GetRuntimeClassName() const
-        {
-            return L"%.%";
-        }
-%    };
-}
-)";
-
-            w.write(format,
-                type_namespace,
-                type_name,
-                bind<write_component_factory_interfaces>(factories),
-                type_namespace,
-                type_name,
-                type_namespace,
-                type_name,
-                bind<write_component_forwarders>(factories));
-        }
-
-        if (non_static)
-        {
-            auto format = R"(
-#if defined(WINRT_FORCE_INCLUDE_%_XAML_G_H) || __has_include("%.xaml.g.h")
-#include "%.xaml.g.h"
-#else
-
-namespace winrt::@::implementation
-{
-    template <typename D, typename... I>
-    using %T = %_base<D, I...>;
-}
-
-#endif
-)";
-
-            std::string upper(type_name);
-            std::transform(upper.begin(), upper.end(), upper.begin(), [](char c) {return static_cast<char>(::toupper(c)); });
-
-            auto include_path = get_generated_component_filename(type);
-
-            w.write(format,
-                upper,
-                include_path,
-                include_path,
-                type_namespace,
-                type_name,
-                type_name);
-        }
-    }
-
-    void write_component_base(writer& w, TypeDef const& type)
-    {
-        if (empty(type.InterfaceImpl()))
-        {
-            return;
-        }
-
-        auto type_name = type.TypeName();
-        auto base_type = get_base_class(type);
-
-        if (base_type && settings.filter.includes(base_type))
-        {
-            w.write(" : %T<%, @::implementation::%>",
-                type_name,
-                type_name,
-                base_type.TypeNamespace(),
-                base_type.TypeName());
-        }
-        else
-        {
-            w.write(" : %T<%>", type_name, type_name);
-        }
-    }
-
-    void write_component_member_declarations(writer& w, TypeDef const& type)
-    {
-        auto type_name = type.TypeName();
-
-        for (auto&& factory : get_factories(type))
-        {
-            if (factory.activatable)
-            {
-                if (!factory.type)
-                {
-                    continue;
-                }
-
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-
-                    w.write("        %(%);\n",
-                        type_name,
-                        bind<write_implementation_params>(signature));
-                }
-            }
-            else if (factory.statics)
-            {
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-                    w.async_types = is_async(method, signature);
-                    auto method_name = get_name(method);
-
-                    w.write("        static % %(%)%;\n",
-                        signature.return_signature(),
-                        method_name,
-                        bind<write_implementation_params>(signature),
-                        is_noexcept(method) ? " noexcept" : "");
-                }
-            }
-        }
-
-        for (auto&&[interface_name, info] : get_interfaces(w, type))
-        {
-            if (info.base)
-            {
-                continue;
-            }
-
-            w.generic_param_stack.insert(w.generic_param_stack.end(), info.generic_param_stack.begin(), info.generic_param_stack.end());
-
-            for (auto&& method : info.type.MethodList())
-            {
-                method_signature signature{ method };
-                w.async_types = is_async(method, signature);
-                auto method_name = get_name(method);
-
-                w.write("        % %(%)%;\n",
-                    signature.return_signature(),
-                    method_name,
-                    bind<write_implementation_params>(signature),
-                    is_noexcept(method) ? " noexcept" : "");
-            }
-
-            w.generic_param_stack.resize(w.generic_param_stack.size() - info.generic_param_stack.size());
-        }
-    }
-
-    void write_component_h(writer& w, TypeDef const& type)
-    {
-        auto type_name = type.TypeName();
-        auto type_namespace = type.TypeNamespace();
-        auto base_type = get_base_class(type);
-        std::string base_include;
-
-        if (base_type)
-        {
-            if (settings.filter.includes(base_type))
-            {
-                base_include = "#include \"" + get_generated_component_filename(base_type) + ".h\"\n";
-            }
-        }
-
-        {
-            auto format = R"(#include "%.g.h"
-%
-namespace winrt::@::implementation
-{
-    struct %%
-    {
-        %() = default;
-
-%    };
-}
-)";
-
-            w.write(format,
-                get_generated_component_filename(type),
-                base_include,
-                type_namespace,
-                type_name,
-                bind<write_component_base>(type),
-                type_name,
-                bind<write_component_member_declarations>(type));
-
-        }
-
-        if (has_factory_members(type))
-        {
-            auto format = R"(namespace winrt::@::factory_implementation
-{
-    struct % : %T<%, implementation::%>
-    {
-    };
-}
-)";
-            w.write(format,
-                type_namespace,
-                type_name,
-                type_name,
-                type_name,
-                type_name);
-        }
-    }
-
-    void write_component_member_definitions(writer& w, TypeDef const& type)
-    {
-        auto type_name = type.TypeName();
-
-        for (auto&& factory : get_factories(type))
-        {
-            if (factory.activatable)
-            {
-                if (!factory.type)
-                {
-                    continue;
-                }
-
-                auto format = R"(    %::%(%)
-    {
-        throw hresult_not_implemented();
-    }
-)";
-
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-
-                    w.write(format,
-                        type_name,
-                        type_name,
-                        bind<write_implementation_params>(signature));
-                }
-            }
-            else if (factory.statics)
-            {
-                auto format = R"(    % %::%(%)%
-    {
-        throw hresult_not_implemented();
-    }
-)";
-
-                for (auto&& method : factory.type.MethodList())
-                {
-                    method_signature signature{ method };
-                    w.async_types = is_async(method, signature);
-                    auto method_name = get_name(method);
-
-                    w.write(format,
-                        signature.return_signature(),
-                        type_name,
-                        method_name,
-                        bind<write_implementation_params>(signature),
-                        is_noexcept(method) ? " noexcept" : "");
-                }
-            }
-        }
-
-        for (auto&&[interface_name, info] : get_interfaces(w, type))
-        {
-            if (info.base)
-            {
-                continue;
-            }
-
-            w.generic_param_stack.insert(w.generic_param_stack.end(), info.generic_param_stack.begin(), info.generic_param_stack.end());
-
-            for (auto&& method : info.type.MethodList())
-            {
-                auto format = R"(    % %::%(%)%
-    {
-        throw hresult_not_implemented();
-    }
-)";
-
-                method_signature signature{ method };
-                w.async_types = is_async(method, signature);
-                auto method_name = get_name(method);
-
-                w.write(format,
-                    signature.return_signature(),
-                    type_name,
-                    method_name,
-                    bind<write_implementation_params>(signature),
-                    is_noexcept(method) ? " noexcept" : "");
-            }
-
-            w.generic_param_stack.resize(w.generic_param_stack.size() - info.generic_param_stack.size());
-        }
-    }
-
-    void write_component_cpp(writer& w, TypeDef const& type)
-    {
-        auto format = R"(#include "%.h"
-
-namespace winrt::@::implementation
-{
-%}
-)";
-
-        w.write(format,
-            get_component_filename(type),
-            type.TypeNamespace(),
-            bind<write_component_member_definitions>(type));
     }
 }
